feat: max-num-threads: setting max number of threads per exploration
It is set at the beginning of the main function, to 2 * NUM_CORES,
where NUM_CORES is the number of cpu cores (not only physical cores).
This needs unsafe code, since it requires a mutable static variable,
which is written to only once, before any reading is done, so that's ok.